% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signature_scores.R
\name{calculate_signature_scores}
\alias{calculate_signature_scores}
\title{calculate signature scores}
\usage{
calculate_signature_scores(
  file,
  geneset,
  info.name = NA,
  info.type = NA,
  plot.title = "Signature scores"
)
}
\arguments{
\item{file}{Filepath/filename of expression data matrix}

\item{geneset}{geneset vector}

\item{info.name}{sample info for plotting if needed}

\item{info.type}{sample info for plotting if needed}

\item{plot.title}{plot title}
}
\description{
Reads file with samples in columns and variables in rows, and geneset list.
Calculates sum z-score signatures across provided samples, and writes to file
Returns dataframe of signature scores
}
