% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_all.R
\name{intersect_all}
\alias{intersect_all}
\title{Intersects all lists}
\usage{
intersect_all(a, b, ...)
}
\arguments{
\item{a, b...}{all lists to intersect}
}
\value{
Returns a vector containing common items
}
\description{
Intersects all lists
}
