% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder plot with ks p-value.R
\name{ladder.plot}
\alias{ladder.plot}
\title{ladder plot with ks p-value}
\usage{
ladder.plot(z, title, comp, cex = 1.5)
}
\arguments{
\item{z}{}

\item{title}{}

\item{comp}{}

\item{cex}{}
}
\value{

}
\description{
ladder plot with ks p-value
}
\examples{

}
