% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder plot with ks p-value.R
\name{ladder.plot.with.transparent.lines}
\alias{ladder.plot.with.transparent.lines}
\title{ladder.plot.with.transparent.lines}
\usage{
ladder.plot.with.transparent.lines(z, title, metric, ladder_color, cex = 1.5)
}
\arguments{
\item{z}{dataframe with list of genes (or similar) and metrics that can be used to rank the genes, 
and indication of the geneset membership based on the gene color (column with colname="color"; 
any color = member, "trnasparent" = non-member)}

\item{title}{title for the plot, typically the name/description of the gene set used for the 
enrichment analysis}

\item{metric}{parameter indicating the metric (column name) in dataframe z to be used to rank genes}

\item{cex}{character enhancement factor - scales the foint size}
}
\value{
RVAL
}
\description{
ladder.plot.with.transparent.lines
}
\examples{
provided below the functions, at the end of the file

}
