% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_from_file.R
\name{PCA_from_file}
\alias{PCA_from_file}
\title{PCA from file}
\usage{
PCA_from_file(file, center = TRUE, scale = FALSE, fread = FALSE,
  tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{file}{Filepath/filename of data matrix}

\item{center}{default=T}

\item{scale}{default=F}

\item{fread}{default=F, use fread for large input files}

\item{tol}{number specifying magnitude below which comps should be ommited}
}
\description{
Reads file with samples in columns and variables in rows, and does PCA. Writes to file scores, loadings, eigenvalues.
}
