% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cca_loadings.R
\name{plot_cca_loadings}
\alias{plot_cca_loadings}
\title{Title: Plot regularized canonical correlation analysis loadings
edited from Estelle's https://github.com/estelleyao0530/Canonical_Correlation_Function}
\usage{
plot_cca_loadings(projx, projy, modulecolor, modulecolor2, number_plot, name)
}
\arguments{
\item{projx}{file, from run_cca output; matrix A loadings}

\item{projy}{file, from run_cca output; matrix B loadings}

\item{modulecolor}{file, wgcna module; variable cluster specification for matrix A}

\item{modulecolor2}{file, wgcna module; variable cluster specification for matrix B}

\item{number_plot}{int, numbers of component to plot}

\item{name}{string, name of output files

import ggplots
import ggpubr
import ggrepel}
}
\description{
Input
Projected loading score from matrix A & B run.cca output
WGCNA modules for matrix A & B
}
\details{
Output:
Scatter plots from specified number of components for matrix A
Scatter plots from specified number of components for matrix B

Note:
1) projected loading avoids collinearity of loadings
2) variables (drugs, genes) are colored with wgcna modules; a two column dataframe - variables in col1, modulecolor in col2
}
