% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_3d.R
\name{plot_pca_3d}
\alias{plot_pca_3d}
\title{plot_pca_3d}
\usage{
plot_pca_3d(
  scores,
  info,
  info.Group = NA,
  PCx = "PC1",
  PCy = "PC2",
  PCz = "PC3",
  indiv_labels = FALSE,
  grouplabels = FALSE,
  Title = "3D PCA Plot",
  drawshape = F
)
}
\arguments{
\item{scores}{Required: dataframe of PCA scores. First column is named "Score", the following columns are "PC1", "PC2",...}

\item{info}{Required: dataframe. First column is named "cellline" and second column contains group info (if there is any).}

\item{info.Group}{Optional. Can run script without any group info.}

\item{PCx}{PC on X axis. Default is "PC1".}

\item{PCy}{PC on Y axis. Default is "PC2".}

\item{PCz}{PC on Z axis. Default is "PC3".}

\item{indiv_labels}{Boolean T/F. Do you want each point to show its corresponding cell line name? Default is FALSE.}

\item{grouplabels}{Boolean T/F. Do you want each grouping name to show on the plot? Default is FALSE.}

\item{Title}{Title name that will display at the top of the plot.}

\item{drawshape}{Boolean T/F. Finds the center 3d point of all points for each group and then connects the dots to form a 3d shape if one exists.}
}
\description{
Used to visualize principle component analysis in 3 dimensional space.
}
\examples{
#' data(iris)
iris.pca <- prcomp(iris[,-c(5)], center = TRUE,scale. = TRUE)
scores<-as.data.frame(iris.pca$x)
scores<-cbind(rownames(iris),scores)
names(scores)[1]<-'Score'
info<-cbind(rownames(iris),as.data.frame(iris[,'Species']))
names(info)[1]<-'cellline'
names(info)[2]<-'Group'

plot_pca_3d(scores=scores,info=info,info.Group=info$Group,grouplabels=TRUE,Title ="3D Plot-Iris Species",drawshape=T)
plot_pca_3d(scores=scores,info=info,info.Group=info$Group,grouplabels=TRUE,Title ="3D Plot-Iris Species")
plot_pca_3d(scores=scores,info=info,indiv_labels=TRUE,Title ="3D Plot-Iris Species")

}
\author{
Alexzandra Morris
}
