% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_paint_gene_repel.R
\name{plot_pca_paint_gene_repel}
\alias{plot_pca_paint_gene_repel}
\title{Plot PCA Paint Gene}
\usage{
plot_pca_paint_gene_repel(
  file,
  gene = "JUN",
  title = "",
  labels = TRUE,
  PCx = "PC1",
  PCy = "PC2",
  ellipse = F,
  conf = 0.95,
  density = F,
  fliph = F,
  flipv = F,
  sDev_file_exists = TRUE
)
}
\arguments{
\item{file}{File containing scores matrix}

\item{title}{Title of the plot}

\item{labels}{default=T}

\item{PCx, PCy}{PCs to display}

\item{ellipse}{Construct confidence region based on groups in info.type, default = T}

\item{conf}{default = 0.95}

\item{density}{plot x-y density plots}

\item{fliph, flipv}{flip plot hoirzontally or vertically}

\item{info.name}{Vector of sample names}

\item{info.type}{Vector of sample types in the same order}
}
\description{
Plots PCA from scores file (output of PCA_from_file)
}
\details{
plot_pca_paint_gene differs from plot_pca in that it uses gradient coloring of points based on the expression values of a gene
plot_pca_paint_gene_repel uses geom_text_repel
}
