% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_projection_labeloptions.R
\name{plot_pca_projection_labeloptions}
\alias{plot_pca_projection_labeloptions}
\title{Plot PCA projection and data points from original PCA}
\usage{
plot_pca_projection_labeloptions(
  file,
  rotated.file,
  info.name,
  info.type,
  info.name2,
  info.type2,
  title = "Projection",
  labelsProj = T,
  labelsOrgPCA = F,
  PCx = "PC1",
  PCy = "PC2",
  ellipse = F,
  conf = 0.95,
  fliph = F,
  flipv = F,
  save = F,
  savename = ""
)
}
\arguments{
\item{file}{scores file of original data}

\item{rotated.file}{File containing scores matrix of projected data}

\item{info.name}{Vector of sample names in original PCA}

\item{info.type}{Vector of sample types in original PCA in the same order as names}

\item{info.name2}{Vector of sample names of projected data}

\item{info.type2}{Vector of sample types of projected data in the same order}

\item{title}{Title of the plot}

\item{labelsProj}{default=T}

\item{labelsOrgPCA}{default=F}

\item{PCx, PCy}{PCs to display}

\item{ellipse}{Construct confidence region based on groups in info.type, default = T}

\item{conf}{default = 0.95}

\item{fliph, flipv}{flip plot hoirzontally or vertically}

\item{save}{save plot as png}

\item{savename}{name of file saved}
}
\description{
Plots projection from rotated.scores file (output of intersect_do_PCA_and_project_second_dataset) on top of original data
}
