% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pls.R
\name{plot_pls}
\alias{plot_pls}
\title{Plot PLS}
\usage{
plot_pls(file, info.name, info.type, title = "", labels = TRUE,
  PCx = "comp.1", PCy = "comp.2", ellipse = F, conf = 0.95,
  fliph = F, flipv = F)
}
\arguments{
\item{file}{File containing scores matrix}

\item{info.name}{Vector of sample names}

\item{info.type}{Vector of sample types in the same order}

\item{title}{Title of the plot}

\item{labels}{default=T}

\item{PCx, PCy}{PCs to display}

\item{ellipse}{Construct confidence region based on groups in info.type, default = T}

\item{conf}{default = 0.95}

\item{fliph}{default = F}

\item{flipv}{default = F}
}
\description{
Plots PLS from scores file (output of PLSR_from_file)
}
