% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_from_file_and_predict_second_dataset.R
\name{PLSR_from_file_and_predict_second_dataset}
\alias{PLSR_from_file_and_predict_second_dataset}
\title{Partial Least Squares and predict second dataset}
\usage{
PLSR_from_file_and_predict_second_dataset(file, file2, sample.names,
  sample.type, y.response, sample.names2 = NULL, sample.type2 = NULL,
  train_string, title = "PLSR", comp.x = "comp.1", comp.y = "comp.2",
  comps = 3, scale = F, labels = F, fread = FALSE)
}
\arguments{
\item{file}{file for X matirx}

\item{file2}{file for test data matrix}

\item{sample.names}{Vector of sample names in X matrix}

\item{sample.type}{vector of sample groups}

\item{y.response}{numeric vector of response values in same order as samples}

\item{sample.names2}{Vector of sample names in 2nd dataset, if needed}

\item{sample.type2}{Vector of sample types in 2nd dataset, if needed}

\item{train_string}{string to insert in file name of predicted scores}

\item{title}{title of the plot}

\item{comp.x, comp.y}{comps to display}

\item{comps}{number of components to compute}

\item{scale}{default=T}

\item{labels}{label the plot, default = T}

\item{fread}{default=F, use fread for large input files}
}
\description{
Builds PLS model from training dataset and predicts second dataset
Writes out predicted.scores of the second dataset
Plots original PLS, projected samples only, and projected samples ontop of original PLS
}
