% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.enrichR.R
\name{run.enrichR}
\alias{run.enrichR}
\title{run.enrichR}
\usage{
run.enrichR(topN)
}
\arguments{
\item{topN}{list of the genes for the enrichmnet analysis}
}
\value{
enricher.df
}
\description{
runs a full enrichR analysis
}
\details{
CRAN
install.packages("enrichR")  
Github
library(devtools)
install_github("wjawaid/enrichR")
}
