% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cca.R
\name{run_cca}
\alias{run_cca}
\title{Title: Run regularized canonical correlation analysis
edited from Estelle's https://github.com/estelleyao0530/Canonical_Correlation_Function
based on http://mixomics.org/methods/rcca/}
\usage{
run_cca(df1, df2, ncomp, save_cca.obj = F, savename)
}
\arguments{
\item{df1}{numeric dataframe/matrix A}

\item{df2}{numeric dataframe/matrix B}

\item{ncomp}{int, numbers of components to output}

\item{save_cca.obj}{logical, default = F; option to save cca object}

\item{savename}{string, name of the output file}
}
\value{
cca object
}
\description{
Input:
matrix A (rows as samples, cols as variables(drugs, genes))
matrix B (rows as samples, cols as variables(drugs, genes))
}
\details{
Note:
1) matrix A and B need to be matched by samples
2) both matrices should be numeric - so sample names should be rownames or removed, not the first column
3) numbers of variables (columns) > numbers of samples (rows)

Output:
saves cca object from mixOmics package (optional)
writes to file average variate scores and projected loadings of decomposed matrices that maximize the correlation between A and B
}
\examples{
run_cca(ctrp, ceres, 9, T, "CCA_Ctrp_Ceres")
}
