% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinning_3d_pca_link.R
\name{spinning_3d_pca_link}
\alias{spinning_3d_pca_link}
\title{Plot PCA scores in 3d space. Link to locally saved 3d html plot gets made and opens automatically in browser after function is run.}
\usage{
spinning_3d_pca_link(
  scores,
  info.type,
  X = PC1,
  Y = PC2,
  Z = PC3,
  title = "3d PCA plot",
  colorlst = NULL,
  marker = NULL
)
}
\arguments{
\item{scores;}{the output from PCA_from_file that has been read in using read.delim - first column includes the samples}

\item{info.type;}{info.type column is in the same order as samples in scores and is a factor. Points will be colored by info.type. It is recommended to make info.type a new column in the scores file.}

\item{X;}{default is PC1}

\item{Y;}{default is PC2}

\item{Z;}{default is PC3}

\item{title;}{title name for 3d plot and the name that is used in saving the html file}

\item{colorlst;}{optional: list of colors that correspond to the unique(info.type)}

\item{marker;}{optional: can use this parameter to adjust size, shape, opacity of points}
}
\value{
html link to 3d Spinning PCA plot
}
\description{
Plot PCA scores in 3d space. Link to locally saved 3d html plot gets made and opens automatically in browser after function is run.
}
\examples{
my_PCA_scores <- read.delim("my_PCA_scores.txt")
my_PCA_scores$type <- info$type[match(my_PCA_scores$Score, info$sample)]
my_PCA_scores$type <- factor(my_PCA_scores$type, levels = unique(my_PCA_scores$type))
spinning_3d_pca_link(scores = my_PCA_scores, info.type = my_PCA_scores$type, X = my_PCA_scores$PC1, Y = my_PCA_scores$PC2, Z = my_PCA_scores$PC3, title = "3d PCA plot", colorlst=c("red", "dodgerblue", "pink", "orange"))
}
