% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimax_from_file.R
\name{varimax_from_file}
\alias{varimax_from_file}
\title{Varimax rotation from file
Last mod: 4/2018, ksheu}
\usage{
varimax_from_file(file.scores, file.loadings, comp = 2, normalize = T)
}
\arguments{
\item{file.scores}{file for scores matrix (ex. output of PCA_from_file or PLSR_from_file)}

\item{file.loadings}{file for scores matrix (ex. output of PCA_from_file or PLSR_from_file)}

\item{comp}{number of components to perform rotation on}

\item{normalize}{default = T Perform Kaiser normalization (rows scaled before normalization, and scaled back afterwards)}
}
\description{
Writes out Varimax rotated loadings and scores
}
