% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder plot with ks p-value.R
\name{ladder.plot}
\alias{ladder.plot}
\title{ladder plot with ks p-value}
\usage{
ladder.plot(z, title, metric, ladder_color, cex = 1.5, latexpdf_flag = 0)
}
\arguments{
\item{z}{dataframe with list of genes (or similar) and metrics that can be used to rank the genes, 
and indication of the geneset membership based on the gene color (column with colname="color"; 
any color = member, "trnasparent" = non-member)}

\item{title}{title for the plot, typically the name/description of the gene set used for the 
enrichment analysis}

\item{metric}{parameter indicating the metric (column name) in dataframe z to be used to rank genes}

\item{cex}{character enhancement factor - scales the font size}

\item{latexpdf_flag}{used to create a formated pdf version of the genes or samples that make up the ladder rungs
only turn latexpdf_flag on if latexpdf is installed and working}
}
\value{
RVAL
}
\description{
ladder plot with ks p-value
}
\examples{
provided below the functions, at the end of the file

}
