% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varimax_gradient_fill.R,
%   R/plot_varimax_gradient_fill_missingGrey90.R
\name{plot_varimax_gradient_fill}
\alias{plot_varimax_gradient_fill}
\title{Plot varimax PCA}
\usage{
plot_varimax_gradient_fill(
  file,
  info.name,
  info.type,
  info.color,
  title = "",
  labels = TRUE,
  PCx = "PC1",
  PCy = "PC2",
  ellipse = F,
  conf = 0.95,
  density = F,
  fliph = F,
  flipv = F
)

plot_varimax_gradient_fill(
  file,
  info.name,
  info.type,
  info.color,
  title = "",
  labels = TRUE,
  PCx = "PC1",
  PCy = "PC2",
  ellipse = F,
  conf = 0.95,
  density = F,
  fliph = F,
  flipv = F
)
}
\arguments{
\item{file}{File containing scores matrix}

\item{info.name}{Vector of sample names}

\item{info.type}{Vector of sample types in the same order}

\item{info.color}{Vector to be used for gradient fill of plot points, with sample types in the same order}

\item{title}{Title of the plot}

\item{labels}{default=T}

\item{PCx, PCy}{PCs to display}

\item{ellipse}{Construct confidence region based on groups in info.type, default = T}

\item{conf}{default = 0.95}

\item{density}{plot x-y density plots}

\item{fliph, flipv}{flip plot hoirzontally or vertically}

\item{missing}{plot, plot.grey90 or do.not.plot points with missing gradient fill values 
(grey90 is lighter than the default grey)}
}
\description{
Plots varimax PCA from scores file (output of PCA_from_file followed by varimax_from_file)

Plots varimax PCA from scores file (output of PCA_from_file followed by varimax_from_file)
}
\details{
plot_varimax_gradient_fill differs from plot_varimax in that it uses gradient coloring of points

plot_varimax_gradient_fill differs from plot_varimax in that it uses gradient coloring of points
}
