% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_from_file.R
\name{PLSR_from_file}
\alias{PLSR_from_file}
\title{Partial Least Squares Regression from file
Last mod: 6/17/18, ksheu}
\usage{
PLSR_from_file(file, sample.names, sample.type, y.response,
  title = "PLSR", comps = 5, scale = F, comp.x = "comp.1",
  comp.y = "comp.2", labels = F, fread = F)
}
\arguments{
\item{file}{file for X matrix}

\item{sample.names}{vector of names of the samples}

\item{sample.type}{vector of sample groups}

\item{y.response}{numeric vector of response values in same order as samples}

\item{title}{title of the plot}

\item{comps}{number of components to compute}

\item{scale}{default=F}

\item{comp.x, comp.y}{comps to display}

\item{labels}{label the plot, default = F}

\item{fread}{defaylt = F}
}
\description{
Writes out Xscores, Xloadings, and PVE.
Plot scores output with plot_pls()
}
